Update and rename main.py to 3func.py
